Read scrollable area height from the element's style
Added code which sets the scrollable area height (o.height) to whatever value is specified in the style of the element to which slimScroll is being added. This is an improvement because it allows different elements of different heights on the same page to use slimScroll with a single initialisation call.